fase 3. separar renderer canvas por capas internas
